content/034/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
httpx==0.27.2
websockets==13.0.1
python-multipart==0.0.12
discord.py==2.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
